third_party/blink/web_tests/platform/win11-arm64/http/tests/eventsource/existent-eventsource-status-error-iframe-crash-expected.txt

CONSOLE ERROR: EventSource's response has a MIME type ("text/html") that is not "text/event-stream". Aborting the connection.
PASS
